protect copy and bugfix content category, and disable user menu

+++ application/config/_function.php
@@ -129,5 +129,8 @@ function recursive_rmdir($directory)
             unlink($file);
         }
     }
-    rmdir($directory);
+    if(is_dir($directory))
+    {
+    	rmdir($directory);
+    }
 }
